refactor: Combine firebase hooks with firebase config files into 1 directory
- and also continue migration of JumpalCommon to .tsx
