+++ .vscode/settings.json
@@ -2,5 +2,11 @@
   "editor.codeActionsOnSave": {
     "source.fixAll": true
   },
-  "eslint.validate": ["javascript", "typescript", "javscriptreact", "typescriptreact"]
+  "eslint.validate": [
+    "javascript",
+    "typescript",
+    "javscriptreact",
+    "typescriptreact"
+  ],
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
