--- apps/frontend/.prettierrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  printWidth: 80,
-};
