--- apps/frontend/src/Firebase/context.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const FirebaseContext = React.createContext(null);
-
-export default FirebaseContext;
