--- apps/frontend/src/Firebase/index.js
@@ -1,6 +1,0 @@
-import FirebaseContext from "./context";
-import Firebase from "./Firebase";
-
-export default Firebase;
-
-export { FirebaseContext };
