+++ apps/frontend/src/data/config.js
@@ -1,0 +1,12 @@
+const config = {
+  apiKey: "AIzaSyB_czfEJ_qq-H58aEHgKV1YyZXc_YzmbAs",
+  authDomain: "jumpal-27f91.firebaseapp.com",
+  databaseURL: "https://jumpal-27f91.firebaseio.com",
+  projectId: "jumpal-27f91",
+  storageBucket: "jumpal-27f91.appspot.com",
+  messagingSenderId: "867715616373",
+  appId: "1:867715616373:web:30087c19543aeacae7b23c",
+  measurementId: "G-T3DR03KQQ3",
+};
+
+export default config;
