+++ apps/frontend/src/data/config.js sample
