+++ apps/frontend/src/data/useAuth.js
@@ -1,6 +1,5 @@
 import { useState, useContext, useEffect } from "react";
-
-import { FirebaseContext } from "../Firebase";
+import { FirebaseContext } from "./Firebase";
 import { onAuthStateChanged } from "firebase/auth";
 
 export default function useAuth() {
