+++ apps/frontend/src/data/useIgDb.js
@@ -1,15 +1,14 @@
 import { useState, useContext, useEffect } from "react";
 import { onAuthStateChanged } from "firebase/auth";
-import { FirebaseContext } from "../Firebase";
-import { getPbTime } from "../utils";
-import { get, set, remove, child, off, onValue } from "firebase/database";
+import { FirebaseContext } from "./Firebase";
+import { get, push, remove, child, off, onValue } from "firebase/database";
 
-export default function usePbDb() {
+export default function useIgDb() {
   const firebase = useContext(FirebaseContext);
   const [user, setAuthUser] = useState(null);
   const [loading, setLoading] = useState(true);
-  const [pb, setPb] = useState([]);
-  const pbRef = firebase.personalBests;
+  const [ig, setIg] = useState([]);
+  const igsRef = firebase.igs;
 
   const unsubscribe = onAuthStateChanged(firebase.auth, (authUser) => {
     if (authUser) {
@@ -19,49 +18,43 @@ export default function usePbDb() {
     }
   });
 
-  const onPbUpdate = (snapshot) => {
+  const onIgUpdate = (snapshot) => {
     setLoading(true);
     if (snapshot) {
-      const newPersonalBests = [];
-      newPersonalBests.push(snapshot.val());
-      setPb(newPersonalBests);
+      const newIg = [];
+      newIg.push(snapshot.val());
+      setIg(newIg);
     }
     setLoading(false);
   };
 
   useEffect(() => {
-    if (onPbUpdate) {
-      onValue(pbRef, onPbUpdate);
+    if (onIgUpdate) {
+      onValue(igsRef, onIgUpdate);
     }
     return () => {
-      off(pbRef);
+      off(igsRef);
       unsubscribe();
     };
   }, []);
 
-  const getPb = async () => {
+  const getIg = async () => {
     try {
       setLoading(true);
       if (user) {
-        const snapshot = await get(pbRef);
-        onPbUpdate(snapshot);
-        return true;
+        const snapshot = await get(igsRef);
+        onIgUpdate(snapshot);
       }
-      return false;
     } catch (e) {
       console.error(e);
       return false;
     }
   };
 
-  const addPb = async (event, score, time) => {
+  const addIg = async (note, url) => {
     try {
       if (user) {
-        const currEventPbRef = child(pbRef, event);
-        await set(currEventPbRef, {
-          score: score,
-          time: getPbTime(time),
-        });
+        push(igsRef, { note, url });
         return true;
       }
       return false;
@@ -71,10 +64,11 @@ export default function usePbDb() {
     }
   };
 
-  const delPb = async (event) => {
+  const delIg = async (id) => {
     try {
       if (user) {
-        await remove(child(pbRef, event));
+        const igRef = child(igsRef, id);
+        await remove(igRef);
         return true;
       }
       return false;
@@ -84,5 +78,5 @@ export default function usePbDb() {
     }
   };
 
-  return [pb, loading, getPb, addPb, delPb];
+  return [ig, loading, getIg, addIg, delIg];
 }
