--- apps/frontend/src/hooks/index.js
@@ -1,7 +1,0 @@
-import useAuth from "./useAuth";
-import usePbDb from "./usePbDb";
-import useSdDb from "./useSdDb";
-import useSlDb from "./useSlDb";
-import useIgDb from "./useIgDb";
-
-export { useAuth, usePbDb, useSdDb, useSlDb, useIgDb };
