+++ apps/frontend/src/index.js
@@ -2,8 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
-import FirebaseContext from "./Firebase/context";
-import Firebase from "./Firebase/Firebase";
+import { FirebaseContext, Firebase } from "./data";
 
 ReactDOM.render(
   <React.StrictMode>
